app.models.light: Add `__init__` method
The method was added to properly handle `bool` types from client
requests. These are received as `str` types in the `request.form` dict
and cannot simply be converted using `bool(value)`. For example:

  >>> bool('False')
  True

will bite you in a bad place. The tests have been updated to account for
the fact that we're now setting the `name` field in `__init__`, which
may raise a `ValidationError` before a `DataIntegrityError` gets a
chance to be raised. (Validators are auto-run on assignment.)

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>