Add Flask-Migrate and database model migration
To initialize, generate, and apply migrations use the following command
sequence:

  $ flask db init
  ...
  $ flask db migrate -m 'Add Light model'
  ...
  $ flask db upgrade
  ...

The `init` command creates the `migrations/` directory and structure.
The `migrate` command generates the migration script and it often needs
to be modified manually. For example, it often fails to include database
constraints (e.g. `sa.CheckConstraint` had to be added manually by me to
validate the length of the `name` field value.)

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>