app.models.light: Improve power state handling and tests
The test cases make sure that we're processing the truthy/falsey values
correctly in the model. In addition, this commit adds handling of the
`StatementError` raised by SQLAlchemy when one of the unexpected falsey
values is encountered. From the client's point of view, this is a
`DataIntegrityError` and it's exposed as such.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>