app.apis.v0.light: Add `delete` method docstring
This documentation should help most readers navigate through some of the
nonsense some people post about how some HTTP requests should be
handled. At the end of the day, it's the standard that matters, not a
personal opinion. (I guess you could argue that giving priority to the
standard is also a personal opinion, and so on, until we raise
`RecursionError`, but that'd be too meta for a Git commit.)

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>