docker/db.Dockerfile: Add Dockerfile and init script
The `db.Dockerfile` defines the image for the database container and the
init script creates the app-specific database, role, and privileges for
its use.

The script is only executed on container startup and only when the
file system path for the database, i.e. `/var/lib/postgresql/data`, is
empty. Subsequent container launches do not cause the scripts to be
re-run.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>