conf/db/pg_hba.conf: Add Postgres auth file
This should've been included in a previous commmit, specifically on:
  b53d740aeccf37726f80fcccfc045d76fe2eb10d

This file, specifically the last line, requires that any direct
connection to the database be done over SSL. Otherwise, it gets rejected
by the server.

When using `pgAdmin4`, be sure to go into the connection properties and
change `SSL` from `Prefer` to `Require`.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>