README.md: Improve usage examples
Added the `-H` option to always tell the server that we're requesting
`application/json` as the `Content-Type`. In addition to being the
expected practice, this also auto-formats the response without having to
redirect the output to `python3 -m json.tool`, which is unable to
process said output if it includes `-i` for HTTP headers.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>