Rename `api` service to `web`
The `web` service will provide an implementation for both the API and
the GUI. The main reason for this can be summarized as follows:

  * A separate `gui` image/container would mostly be a duplicate of `api`
    because it's only a subset of the existing python code to serve HTML
    and other static content.

  * The `gui` wouldn't be able to generate URLs for `api` service
    endpoints with `flask.url_for` because those endpoints wouldn't be
    registered in it and, therefore, wouldn't be visible.

In other words, for `flask.url_for` to work in a separate `gui` service,
it would need to also host a duplicate/copy of the `api` to register the
API endpoints. In order to avoid this, the originally separate `gui` and
`api` services will be merged into a single `web` service.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>