Add GUI home page, bootstrap4, and dependencies
This commit adds a working main/home page for the web browser. The GUI
is based on Bootstrap4, along with its CSS and JavaScript. The resources
included are minimized.

The HTML files are based on Jinja2 templates, which have many features
over plain hard-coded HTML. Some of these include inheritance/extension
based on blocks to reduce duplication and calling Flask/Python functions
and/or filters (e.g. `url_for`) to generate URLs, among others.

Added the `app.guis` module to process the new GUI routes and endpoints.
The routes are also registered on `app.__init__` for them to be
accessible.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>