lights/light_list.html: Add lights list frontend page
The HTML page contains JavaScript that queries the RESTful API for the
data using AJAX. For this to work, the slim JQuery module was replaced
by the standard module to include support. (Otherwise, it errors out
saying `$.ajax` is not a function.

Added a new `{% block javascript %}` section to the base.html file
following the JavaScript definitions. This is because the whole page
needs to be loaded for some JQuery functionality to become available and
defined (e.g. `ajax` functions).

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>