lights/light_list.html: Refactor to simpler non-JS impl.
After the first JS-driven implementation, I preferred the simpler non-JS
version I originally had, but that never committed. Since there's
already a JS-driven implementation for the detail view, this one should
work fine. Still, there's a record of what it used to be, in case anyone
wants to `git checkout` that commit and take a look.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>