app.services.light: Refactor functions to be kwargs-based
The `get_light` function was ID-based, but that's too limited. It now
takes a set of kwargs that get AND-ed by the database backend in a query.

This should allow finding a single object based on different criteria,
such as finding it by name or some other field. Naturally, this introduces
ambiguity b/c now multiple results can be found.

This function specifically enforces a single result returned policy and
will raise a new `UniqueObjectExpectedError` to let the callers know about
their mishaps. It's the client's responsibility to narrow down the search
criteria enough to uniquely identify the instance they're interested in.
If they cannot do that, then they should use a function or endpoint that
returns a list.

Another change is for `get_light_list`. The `filters` argument are now
`**kwargs` instead of a single dictionary. It makes the public interface
of the function nicer to use.

This commit includes changes for all dependencies (e.g. unit tests,
views, etc).

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>