Add support for Postgres schemas
This commit adds support for organizing database objects within schemas,
including a SQLite-specific workaround to fix issues with unit tests.
The main database used by this system is Postgres, but the automated test
suite makes use of an in-memory SQLite database, for performance reasons.

After refactoring models to use Postgres schemas[1], tests start to fail
b/c SQLite is unable to find the tables it had created in its in-memory
database.

Since Postgres cannot run as an in-memory DB to replace SQLite for tests
and SQLite does not support Postgres schemas[1], a workaround was needed.

This workaround uses the `ATTACH DATABASE ':memory:' AS <schema>` query
to create multiple in-memory SQLite databases that map to the expected
Postgres schemas, allowing queries with the `<schema>.<table>` format to
work by finding tables as expected.

Not doing this causes SQLite to raise errors claiming it cannot find a
database with the schema name. For example:

  sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) unknown database "schema1"
  [SQL: PRAGMA "schema1".table_info("table1")]

This commit includes a fix for this issue.

The `tests.utils` module needed updates to most of the `setup_<object>`
functions to make sure they search the schemas where the tables exist.
Otherwise, they fail to find their tables when checking for duplicates
and end up trying to re-create them every time, which causes errors
about duplicate tables.

[1] The term "schema" means different things for Postgres and SQLite. For
    SQLite, it's a synonym for "database", while for Postgres, it's an
    an object contained within a database (e.g. to organize things, used
    as namespaces) with its own properties, privileges, etc.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>