Add `date_created` field, tests, migrations, etc
This commit adds a new `date_created` field for the `Light` model as
well as all of its dependencies, including model and API unit tests,
database migrations, and JSON schema changes.

The best-practice for dealing with date/time data is to always store it
as UTC *without* timezone information. Thus, this is what we do in this
app. The reason behind this is that it is the *server*, not the client,
who manages this piece of data, but the clients may be anywhere in the
world, at least in principle. In addition, the app does not need to
concern itself with the consequences and implications of daylight-saving
changes (i.e. DST) in different parts of the world or even across the USA
itself.

This helps keep the server-side simpler and avoids other complications
that may arise, including security vulnerabilities. For example, a JSON
Web Token (JWT) with an expiration date could remain valid for a longer
period of time, thus continuing to grant access to someone who should no
longer have any.

To solve this, date/time and zone convertions are always performed on
the client-side using the client's own timezone info. This conversion is
performed via MomentJS, a JS library. The data gets formatted for
MomentJS by the server using Babel.

Signed-off-by: code-dredd <ray.l.rivera@gmail.com>